Platforms/ARM/Juno: Use PCDs for Acpi DBG2
The DBG2 table is used to publish the presence of debug ports available
on a system.

The original DBG2 used explicit constants for hardware details.
That made it difficult to readilly adjust the use of serial ports.
This change modifies the explicit constants to use PCDs, so the serial
port usage can be adjusted at build.  This can help avoid problems when
sharing ports (e.g. between UEFI trace and remote debug) causes fails.

This patch also changes the UART port used to Juno UART1. This enables
the use of a remote Debugger when the EMS terminal is active (UART0).

Dependency:
* This patch requires ARM Trusted Firmware to configure UART1 for
  Non-Secure World access.

Change-Id: I7a9f3c600a13156a7bb450cf45ade336ecbe923c
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Evan Lloyd <evan.lloyd@arm.com>