Platforms/ARM/Juno: Use PCDs for Acpi DBG2
The DBG2 table is used to publish the presence of debug ports available
on a system.  The original DBG2 used explicit constants for hardware
details.  That made it difficult to readilly adjust the use of serial
ports.  This change modifies the explicit constants to use PCDs, so the
serial port usage can be adjusted at build.  This can help avoid
problems when sharing ports (e.g. between UEFI trace and remote debug)
causes fails.

NOTE: Because this change does not modify the UART in use, the existing
      problems with serial debug are still present.  Changing to UART1
      is possible, but requires an update to the ARM Trusted Firmware to
      allow non-secure use of that UART.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Evan Lloyd <evan.lloyd@arm.com>