Platforms/AMD/Overdrive: fix GIC MMIO region sizes
As reported by Punit, Xen panics in the following way when booted on
Overdrive when using this firmware:

(XEN) ****************************************
(XEN) Panic on CPU 0:
(XEN) GICv2: Sizes of GICC (0x00000000002000) and GICV (0x00000000010000) don't match
(XEN)
(XEN) ****************************************

This is due to the fact that the MMIO region sizes deviate from the
Linux upstream version of the DTS for no good reason. So fix that.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>
Acked-by: Punit Agrawal <punit.agrawal@arm.com>