Platforms/AMD/Styx: remove incorrect timer frequency
The ARM generic timer needs to be programmed with the actual frequency
of the input clock, but this can only be done from the most privileged
execution mode implemented by the hardware. UEFI on AArch64 usually
executes in EL2, which is not the most privileged execution mode in most
cases, and so the timer driver is set up to deal with this: no attempt
is made to program the PCD value PcdArmArchTimerFreqInHz into the frequency
register. However, a non-zero PCD value is still treated as an override
for the register value, in case the programmed value is known to be
incorrect.

However, on the various Styx based platforms, the PCD value is set to
an incorrect non-zero value, and so the routines that convert time delays
into cycle counts are off by 33% (187.5 MHz vs 250 MHz). This may affect
timeouts related to SATA link training, and other low level routines that
rely on accurate timekeeping.

So remove the explicit PCD settings, so they default to 0, letting the
driver use the programmed value instead.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>