Platforms/AMD/Styx: set SATA port mode to Gen3 on all ports
The SATA related PCDs consumed by AmdSataInitLib contain a PcdSataPortMode
PCD that sets the port mode to all ports. This PCD defaults to zero, while
the code in question has no default, i.e., the value 0 is not treated as
either Gen1, Gen2 or Gen3, and so the init sequence is not carried out
correctly.

While observed SATA link detection issues on CelloBoard appear to be
unrelated to this (i.e., this change did not improve the situation),
let's set the correct values nonetheless.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>