Platforms/AMD/StyxDtbLoaderLib: disable SMMUs for absent hardware
The logic regarding how the various SMMUs are exposed in the device tree
is inverted, in the sense that they are present in the static DTB image,
and are removed if no SMMU support is requested.

However, the logic is flawed in the sense that it did not remove SMMUs
for hardware that is not there to begin with, i.e., the XGBE network
ports on Cello/Softiron 1000 or the second SATA controller on B1 silicon.

So fix that.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>