Platforms/AMD/Styx: add experimental command line flash tool
This implements a standalone UEFI Shell application that updates the
EFI NOR partition with the contents of STYX_EFI.Fv. Note that this means
that
a) the ATF partition and EFI variables are preserved, *but*
b) this updater can only be used on systems that have already been flashed
   with a firmware build based on this open source branch (and not on systems
   still running AMI firmware)

NOTE: This tool performs no verification whatsoever whether the platform it
      is being executed on is compatible with the image it attempts to flash,
      nor does it check whether the flash completed successfully. This means
      you should only use it if you have a means of unbricking the board if
      needed, such as a Dediprog SPI flasher. You have been warned!

In order for the flasher to have access to the various PCDs that describe
where STYX_EFI.Fv lives in the NOR, it needs to be built as part of the
platform, but after STYX_EFI.Fv has been generated, which results in a
chicken-and-egg situation. Therefore, the recommended way of generating
the flasher is

1) delete the Build/<platform> directory entirely
2) build the platform
3) build the platform again, but append the following

  -m OpenPlatformPkg/Platforms/AMD/Styx/Applications/StyxFlashUefi/StyxFlashUefi.inf \
  -D DO_FLASHER=TRUE

The flasher application is called 'StyxFlashUefi'

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Tested-by: Roy Franz <roy.franz@cavium.com>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>