Platforms/AMD/Overdrive: disable second SATA port
The second SATA port is only available on B1 silicon, but as it turns
out, not on /all/ versions of it: the Overdrive 3000 is B1 based as well,
but any attempts to use the second SATA controller cause the firmware to
crash. So just disable it.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>